Separando o controller cursos do controller video e separando no resource também
