+++ resources/views/cursos/videos/video.blade.php
